chore: Run clippy --fix

+++ crates/buttplug_server/src/device/protocol_impl/vorze_sa/piston.rs
@@ -57,13 +57,13 @@ impl ProtocolHandler for VorzeSAPiston {
 
     self
       .previous_position
-      .store(position as u8, Ordering::Relaxed);
+      .store(position, Ordering::Relaxed);
 
     Ok(vec![
       HardwareWriteCmd::new(
         &[feature_id],
         Endpoint::Tx,
-        vec![VorzeDevice::Piston as u8, position as u8, speed],
+        vec![VorzeDevice::Piston as u8, position, speed],
         true,
       )
       .into(),
